抽取切换页面的逻辑为 hook
